Remove kind of circular RPC event
When the host frame sends a `focusAnnotations` RPC event to the guest
frame, we want to relay the event to the sidebar frame.

However, when the `focusAnnoations` RPC event is coming from the sidebar
frame, we don't want the guest frame to send the event back to the
sidebar frame.

[Based in this
comment](https://github.com/hypothesis/client/pull/4128#issuecomment-1017804981)