WIP - Limit VitalSource iframe height in PDFs
VitalSource PDF books currently create very tall iframes. When a user
scrolls through a PDF page, they are scrolling the _parent_ frame rather
than the iframe itself. The bucket bar and scrolling logic in the client
don't support this. The result is that:

 - Bucket bar buckets are in the wrong location and don't scroll as the
   user scrolls the parent frame
 - Clicking on an annotation / bucket tries to scroll only the iframe,
   not the parent frame.

To resolve this we either need to 1) Support these tall iframes in the
client generally, 2) Forcibly change the height of the iframe, or 3) Work with
VitalSource to change the page layout - they previously did this for
EPUBs already. This commit implements a version of (2).